Create task1.java
Top-Selling Item Check
The store tracks which item was sold the most during a day. Given a list of transactions (each consisting of an item ID and quantity sold), write a program to identify the item ID with the highest total quantity sold. All item IDs are between 5001 and 5005. Return both the item ID and quantity sold. Apply a 20% discount for items 5004 and 5005 only during billing, but not in the comparison. For example, from transactions [(5001, 4), (5004, 6), (5001, 3)], item 5004 should be reported as the top-selling item.